Move robo test to new module
Move custom robo runner to new module
Add robo test to new module
