+++ robotests/build.gradle
@@ -1,0 +1,23 @@
+buildscript {
+    repositories {
+        jcenter()
+    }
+
+    dependencies {
+        classpath 'com.novoda:gradle-android-test-plugin:0.10.1'
+    }
+}
+
+apply plugin: 'java'
+apply plugin: 'android-test'
+
+android {
+    projectUnderTest ':app'
+}
+
+dependencies {
+    testCompile 'junit:junit:4.11'
+    testCompile 'org.mockito:mockito-core:1.9.5'
+    testCompile 'com.squareup:fest-android:1.0.+'
+    testCompile 'org.robolectric:robolectric:2.1.+'
+}
