+++ robotests/src/main/java/com/example/MainActivityRoboTest.java
@@ -1,6 +1,9 @@
-package com.genymobile.testit;
+package com.example;
 
 import android.view.View;
+
+import com.genymobile.testit.MainActivity;
+import com.genymobile.testit.R;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -10,14 +13,13 @@ import static junit.framework.Assert.assertEquals;
 
 @RunWith(CustomRobolectricTestRunner.class)
 public class MainActivityRoboTest {
-
     @Test
     public void testWhenActivityCreatedBatteryLevelViewIsVisible() throws Exception {
+
         MainActivity activity = new MainActivity();
-
         ActivityController.of(activity).attach().create();
-
         int visibility = activity.findViewById(R.id.battery_level_view).getVisibility();
+
         assertEquals(visibility, View.VISIBLE);
     }
 }
